# api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/audit.list": {
      "get": {
        "operationId": "AuditController_list",
        "parameters": [
          {
            "name": "workspace_id",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "user_id",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "action",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "target_type",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "jwt-auth": []
          }
        ],
        "summary": "List audit logs with optional filters",
        "tags": [
          "audit"
        ]
      }
    },
    "/api/audit.getByTarget": {
      "get": {
        "operationId": "AuditController_getByTarget",
        "parameters": [
          {
            "name": "target_id",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "jwt-auth": []
          }
        ],
        "summary": "Get audit logs for a specific target",
        "tags": [
          "audit"
        ]
      }
    },
    "/api/apiKeys.create": {
      "post": {
        "operationId": "ApiKeysController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateApiKeyDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "API key created successfully. Returns the full key (only once).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateApiKeyResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "jwt-auth": []
          }
        ],
        "summary": "Create a new API key",
        "tags": [
          "api-keys"
        ]
      }
    },
    "/api/apiKeys.list": {
      "get": {
        "operationId": "ApiKeysController_list",
        "parameters": [
          {
            "name": "user_id",
            "required": false,
            "in": "query",
            "description": "Filter by user ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "workspace_id",
            "required": false,
            "in": "query",
            "description": "Filter by workspace ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "Filter by status",
            "schema": {
              "enum": [
                "active",
                "revoked",
                "expired"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "jwt-auth": []
          }
        ],
        "summary": "List API keys with optional filters",
        "tags": [
          "api-keys"
        ]
      }
    },
    "/api/apiKeys.get": {
      "get": {
        "operationId": "ApiKeysController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "query",
            "description": "API key ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "jwt-auth": []
          }
        ],
        "summary": "Get API key by ID",
        "tags": [
          "api-keys"
        ]
      }
    },
    "/api/apiKeys.revoke": {
      "post": {
        "operationId": "ApiKeysController_revoke",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "string",
                    "description": "API key ID to revoke"
                  },
                  "revoked_by": {
                    "type": "string",
                    "description": "User ID who is revoking the key"
                  }
                },
                "required": [
                  "id",
                  "revoked_by"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "jwt-auth": []
          }
        ],
        "summary": "Revoke an API key",
        "tags": [
          "api-keys"
        ]
      }
    },
...
    "/api/analytics.query": {
      "post": {
        "operationId": "AnalyticsController_query",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalyticsQueryDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "jwt-auth": []
          }
        ],
        "summary": "Execute an analytics query",
        "tags": [
          "analytics"
        ]
      }
    },
    "/api/analytics.extremes": {
      "post": {
        "operationId": "AnalyticsController_extremes",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExtremesQueryDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "jwt-auth": []
          }
        ],
        "summary": "Get min/max extremes of a metric across grouped data",
        "tags": [
          "analytics"
        ]
      }
    },
...
    "/api/analytics.dimensions": {
      "get": {
        "operationId": "AnalyticsController_getDimensions",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "jwt-auth": []
          }
        ],
        "summary": "Get available dimensions",
        "tags": [
          "analytics"
        ]
      }
    },
...
  },
  "info": {
    "title": "Staminads API",
    "description": "Web analytics platform for tracking TimeScore metrics",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "jwt-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Enter JWT token",
        "in": "header"
      },
      "demo-secret": {
        "type": "apiKey",
        "in": "query",
        "name": "secret",
        "description": "Demo secret for demo endpoints"
      }
    },
    "schemas": {
...
      "CreateApiKeyDto": {
        "type": "object",
        "properties": {
          "user_id": {
            "type": "string"
          },
          "workspace_id": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100
          },
          "description": {
            "type": "string",
            "maxLength": 500
          },
          "scopes": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "expires_at": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "user_id",
          "name",
          "scopes"
        ]
      },
      "CreateApiKeyResponseDto": {
        "type": "object",
        "properties": {
          "key": {
            "type": "string",
            "description": "The full API key. This is ONLY returned once and is not stored.\nFormat: sk_live_<64 hex chars>"
          },
          "apiKey": {
            "type": "object",
            "description": "Public API key metadata (without the key_hash)"
          }
        },
        "required": [
          "key",
          "apiKey"
        ]
      },
...
      "FilterDto": {
        "type": "object",
        "properties": {
          "dimension": {
            "type": "string"
          },
          "operator": {
            "type": "string",
            "enum": [
              "equals",
              "notEquals",
              "in",
              "notIn",
              "contains",
              "notContains",
              "gt",
              "gte",
              "lt",
              "lte",
              "isNull",
              "isNotNull",
              "between",
              "isEmpty",
              "isNotEmpty"
            ]
          },
          "values": {
            "type": "array",
            "items": {
              "type": "object"
            }
          }
        },
        "required": [
          "dimension",
          "operator"
        ]
      },
      "DateRangeDto": {
        "type": "object",
        "properties": {
          "start": {
            "type": "string"
          },
          "end": {
            "type": "string"
          },
          "preset": {
            "type": "string",
            "enum": [
              "last_30_minutes",
              "today",
              "yesterday",
              "last_7_days",
              "last_14_days",
              "last_28_days",
              "last_30_days",
              "last_90_days",
              "last_91_days",
              "this_week",
              "last_week",
              "this_month",
              "last_month",
              "this_quarter",
              "last_quarter",
              "this_year",
              "last_year",
              "last_12_months",
              "all_time"
            ]
          },
          "granularity": {
            "type": "string",
            "enum": [
              "hour",
              "day",
              "week",
              "month",
              "year"
            ]
          }
        }
      },
      "AnalyticsQueryDto": {
        "type": "object",
        "properties": {
          "workspace_id": {
            "type": "string"
          },
          "metrics": {
            "minItems": 1,
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "dimensions": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "filters": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FilterDto"
            }
          },
          "dateRange": {
            "$ref": "#/components/schemas/DateRangeDto"
          },
          "compareDateRange": {
            "$ref": "#/components/schemas/DateRangeDto"
          },
          "timezone": {
            "type": "string"
          },
          "order": {
            "type": "object"
          },
          "limit": {
            "type": "number",
            "minimum": 1,
            "maximum": 10000
          },
          "havingMinSessions": {
            "type": "number",
            "minimum": 1
          }
        },
        "required": [
          "workspace_id",
          "metrics",
          "dateRange"
        ]
      },
      "ExtremesQueryDto": {
        "type": "object",
        "properties": {
          "workspace_id": {
            "type": "string"
          },
          "metric": {
            "type": "string"
          },
          "groupBy": {
            "minItems": 1,
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "dateRange": {
            "$ref": "#/components/schemas/DateRangeDto"
          },
          "filters": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FilterDto"
            }
          },
          "timezone": {
            "type": "string"
          },
          "havingMinSessions": {
            "type": "number",
            "minimum": 1
          }
        },
        "required": [
          "workspace_id",
          "metric",
          "groupBy",
          "dateRange"
        ]
      },
...
    }
  }
}